requirements.txt
numpy
pandas
pillow
scikit-learn
joblib